checks to make sure if an attendee exists

+++ application/controllers/attendee.php
@@ -237,15 +237,24 @@
 		
 		public function delete(){
 			if($_SERVER['REQUEST_METHOD'] == "POST"){
-				$data = array(
+				if($this->attendees->view_where(array("attendeeID" => $this->input->post("inputAttendeeID")))->num_rows()>0){
+					$data = array(
 					"attendeeID" => $this->input->post("inputAttendeeID")
-				);
-				$this->attendees->delete($data);
-				echo json_encode(array(
-						"success" => true,
-						"responseMsg" => "Attendee has been removed!"
-					)
-				);
+					);
+					$this->attendees->delete($data);
+					echo json_encode(array(
+							"success" => true,
+							"responseMsg" => "Attendee has been removed!"
+						)
+					);
+				}
+				else{
+					echo json_encode(array(
+							"success" => false,
+							"responseMsg" => "No such Attendee exists!"
+						)
+					);
+				}
 			}
 			else{
 				show_404();
